FIX: KHO
Fix chức năng quản lý kho (Cập nhật số lượng tồn)

+++ DoAn/QL_HDPHONGLAB/QL_HDPHONGLAB/Model/HOACHAT.cs
@@ -31,9 +31,9 @@ namespace QL_HDPHONGLAB.Model
 
         public int? LUONGNHAP { get; set; }
 
-        public int? LUONGXUAT { get; set; }
+        public double? LUONGXUAT { get; set; }
 
-        public int? LUONGTON { get; set; }
+        public double? LUONGTON { get; set; }
 
         public int? LUONGTHANHLY { get; set; }
